requirements.txt
Django>=4.2 
djangorestframework>=3.12.4
django-rest-swagger==2.2.0
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0
django-cors-headers>=4.2.1
